Addition of the atf-identify utility
Add the atf-identify utility, to be used internally by the atf (either
from the shell interface or from the compiled one).  This calculates the
identifier of a test program in a directory hierarchy.

This is basically (and just for the moment) a proof of concept of how
to write a utility by using all the other classes we have now.  The code
has to be cleaned up (for example, the identify() function does not belong
to the tool; it has to be moved to one of the two libraries to be
available from the binary interface) and reorganized.

Monotone-Parent: dd3b43625ad53d2f95f3775f1e24e4109eb09fa3
Monotone-Revision: 59aca52db7a7e8dab3a2f42444bd55a7ef8b84ec

Monotone-Author: jmmv@NetBSD.org
Monotone-Date: 2007-06-26T21:57:04
Monotone-Branch: org.NetBSD.atf.src