Install the atf test programs
One of the key goals of atf is to allow the installation of test cases in
a production environment, allowing any end user (not only the developers)
to run the tests.  Follow this idea by installing the atf tests in the
system so that the user can easily run them by issuing 'atf-run' in the
appropriate tree.

while doing this, put the units test in an 'atf' subdirectory so that the
build tree can be used in the same way as the installed one.

Monotone-Parent: 3ed9a22201e059c4b662e94253ff3d2d281bb3d0
Monotone-Revision: 580068f39d19e59f92d43622aaab3792ad392d00

Monotone-Author: jmmv@NetBSD.org
Monotone-Date: 2007-06-30T15:44:15
Monotone-Branch: org.NetBSD.atf.src