The test case's function exit code should not be relevant; fix previous
The parent change introduced an heuristic to determine whether the test
case failed unexpectedly (i.e. due to an internal error) or not.  This
was done incorrectly, though, as the exit code of the test case's
function became relevant.  It shouldn't, as otherwise it's easy to get
mysterious unexpected failures for test cases that end with a (test)
command that fails.  Furthermore this was inconsistent with the C++
interface where the test case's bodies do not return any value.

Change this to behave as it should but by ignoring the exit code of the
test case's function.

Monotone-Parent: 7e764448c297de10c501477f25d2318fc83f04b4
Monotone-Revision: fc734907cc28698031da9c44047a30000131bc5e

Monotone-Author: jmmv@NetBSD.org
Monotone-Date: 2007-07-26T07:48:12
Monotone-Branch: org.NetBSD.atf.src