Merge the atfprivate and atf libraries
While having atfprivate and atf separate seemed a good idea at first, it
causes more trouble than desired.  The problem relies in that there are
situations where code needs to use atf functionality from inside atfprivate
but this cannot be done because atfprivate does not depend (and cannot
depend) on atf.  This is specially troublesome as regards exception
classes.

It'd be different if atfprivate was named differently (i.e. atfhelpers) to
refer to the fact that it provides only helper functionality (and not
something that is private and hidden).  But in this case, the library could
have to 1) be installed completely and 2) be put in its own namespace.
Still, the benefits of this are not big.

Just keep things simple until there is a real need for the split (if there
ever is such a need).

Monotone-Parent: 8bf05effdf12f27a3f22111a4b0cabb319e83075
Monotone-Revision: 5f375bdf2c22c72d93d1404c860291572dc6e561

Monotone-Author: jmmv@NetBSD.org
Monotone-Date: 2007-08-12T12:11:31
Monotone-Branch: org.NetBSD.atf.src