Properly skip some test cases if atf-compile is not installed
At the moment, the test suite uses atf-compile in multiple places to
create helper test programs.  This is suboptimal because, under NetBSD,
atf-compile will be part of the comp.tgz set, which may not be installed.

Add a note in the to-do file to address this in the future and properly
skip the affected test cases if the tool is not available.

Monotone-Parent: 4d771b9b2238dc6684a31408b04d1932bd65ffb9
Monotone-Revision: e96d56cd20c567828d17c46dbbf04343e4443fc2

Monotone-Author: jmmv@NetBSD.org
Monotone-Date: 2007-08-16T14:23:51
Monotone-Branch: org.NetBSD.atf.src