Add preliminary manual pages for the programs and formats
Add some manual pages for the ATF programs (tools plus test programs), a
placeholder for documentation of data formats and an introductory document.

Force the installation of distribution documentation as it is referred to
by the introductory manual page.  An alternative approach to this that may
be implemented in the future is to convert these documents to manual pages
and install them accordingly.

This also makes the specification of a manual page name for applications
mandatory, as test programs now have a page to refer to.

Monotone-Parent: 54a76e8ed40b50519ce77264e71124f9680fb578
Monotone-Revision: 225626715f261237fa684de6b8bea1d96b9ed1ed

Monotone-Author: jmmv@NetBSD.org
Monotone-Date: 2007-08-18T12:19:35
Monotone-Branch: org.NetBSD.atf.src