Fix atf-run under Linux
The algorithm that read from a test program's output and muxed its three
descriptors (stdout, stderr and results) was broken under Linux.  I bet
this was a race condition between the buffering introduced by iostreams
and the events read by poll(2).  As the code heavily relied on what poll(2)
reported, buffering broke it.  After thinking about this... I have no
clue why it did *not* break under NetBSD nor Mac OS X.

Change the algorithm to not use poll for the results descriptor, as that
is not needed and simplifies the algorithm a lot.

Use unbuffered I/O to read from stdout and stderr so that poll can safely
be used.  This is probably very inefficient but is the simplest way to
solve it.  (Theoretically we need not use poll to achieve the same results,
but not doing so means that we cannot get immediate reads from the two
descriptors; we have to wait for one of them to finish instead...)

Monotone-Parent: db06672a1ff348cbd0a18d6c67a2a627c29d61f4
Monotone-Revision: 08d78728aa70fac78685aaa3d06935b2f19d1c18

Monotone-Author: jmmv@NetBSD.org
Monotone-Date: 2007-08-20T08:59:08
Monotone-Branch: org.NetBSD.atf.src