Remove the test cases' "isolated" property
The 'isolated' property was added a long time ago to be able to specify
that some test cases did not touch the file system at all.  While this
sounds like a nice idea, manually dealing with this detail is very
error-prone and I'm sure many people will abuse it for micro-optimization
reasons.

Not to mention that the framework already requires to use temporary files
in any case, so the value of 'isolated' is less than it was supposed to
have when it was implemented.

So... simply kill all traces of the 'isolated' property and always enter
a work directory when running a test case.  We get rid of some delicate
and ugly code that had bugs (e.g. incorrect usage of mktemp(3)).

Monotone-Parent: 0bd374619c651525c2ab227ab6a52ef408a31690
Monotone-Revision: 5e6904a7ba00e3f350798e79037c729ca53c03f8

Monotone-Author: jmmv@NetBSD.org
Monotone-Date: 2007-09-30T19:27:17
Monotone-Branch: org.NetBSD.atf.src