Execute the cleanup routine in a subprocess
Change the C++ test programs to run the cleanup routine in a subprocess,
just as the shell ones do.  This is made for safety reasons (so that any
code in the cleanup routines cannot affect other test cases) and to allow
us the removal of chdirs in the parent process later on.

This shouldn't affect performance too much, but if it does, we can always
prevent the execution of the whole cleanup code if the hook is not provided.

Monotone-Parent: 5e6904a7ba00e3f350798e79037c729ca53c03f8
Monotone-Revision: 2b1e13c72040bdad322cbd558cd97677fc690bf1

Monotone-Author: jmmv@NetBSD.org
Monotone-Date: 2007-10-01T16:47:09
Monotone-Branch: org.NetBSD.atf.src