Add a routine to kill a tree of processes
Add a routine, named kill_tree, which can be used to kill a tree of
processes.  This is added to a new module named procs that will contain
stuff to deal with processes.

This will be needed to implement the timeout of test cases.  Given the
complexity of this (specially of its tests), I'm committing it now that
it seems to be working fine under Mac OS X.  Other platforms will probably
still fail somewhere.  Note that some of the new code is ugly, very ugly,
but I don't want to lose it now that it works.

A future step will be the addition of a command-line frontend for this so
that the shell interface can use it to implement timeouts too.

Monotone-Parent: 76a4ec2a9494933bf63e3f98e5ddda1118dc357a
Monotone-Revision: c8c63ddfbff70e116333afeb764f59a6b0c25662

Monotone-Author: jmmv@NetBSD.org
Monotone-Date: 2008-01-14T16:26:21
Monotone-Branch: org.NetBSD.atf.src