Implement a time-limit for test cases
Finally!  Add a new property to test cases, timeout, that can be used to
specify a maximum time-limit for a specific test.  If the test runs over
the maximum allocated time, it will be killed in an attempt to progress
in the execution of the rest of the suite.

Monotone-Parent: 44341041177487716a009378417a4ae211f50477
Monotone-Revision: 30c6286095c07c1450bcc04f1fe0015f227d1a12

Monotone-Author: jmmv@NetBSD.org
Monotone-Date: 2008-01-15T08:43:30
Monotone-Branch: org.NetBSD.atf.src