Move the paragraph formatting algorithms to the C library
Move the algorithms to reformat a set of paragraphs (so that they fill on
screen) to the C library.  To do this, also add a "class" to easily manage
strings in C.

This is still very dirty (ugly code, and completely stupid stuff such as
incomplete C++ tests for C code) but it seems to work and is enough for
today.

Monotone-Parent: ddd495a26be2d473700cce36473f98af162aa9a5
Monotone-Revision: 7591ef751f7191eeaa8fa95d14ae46bf31502ff7

Monotone-Author: jmmv@NetBSD.org
Monotone-Date: 2008-02-10T22:59:01