Add the atf_list_t type
Add a simple linked list implementation as the new atf_list_t type, which
is going to replace uses of sys/queue.h.  The advantage of this new type
over, for example, the TAILQ provided by the system is that lists can be
created dynamically and we not need modify the contents of the objects they
hold to add a "link" entry.  Oh, and it is supposedly more portable.

Monotone-Parent: 189328f9e6589c466377ae71c3dfe049497624d6
Monotone-Revision: fb302de73586c4f7926b535d319a628cfd2cbf89

Monotone-Author: jmmv@NetBSD.org
Monotone-Date: 2008-02-17T12:47:01