Make atf_check_exec be a free function
The "constructor" of atf_check_result is not really a constructor of the
type.  It should be a free function that does the operation and returns
an object of that type, so let be it.

Adjust the C++ interface and the callers to obey this change, too.

Monotone-Parent: d3f8cc48eb9e8a83f2f28aead2cdfaaad0d98db0
Monotone-Revision: 412e4e62e1a81ddbcbef5a61c23d70a15bd9c817

Monotone-Author: jmmv@NetBSD.org
Monotone-Date: 2008-08-26T02:50:01