Convert distribution documents to XML
First attempt at generating the distribution documents from XML versions.
This is still very preliminary, as things may break and markup is not as
good as it could be.

The idea is to be able to publish part of this documentation online in
HTML format.  Of special interest is the list of changes between versions,
which is currently duplicated on the web site's code.

I am not convinced that keeping pregenerated copies of this documentation
in the repository is a good idea... but this looks like a must.  If someone
checks out the code from the repository, we want them to be able to read
the installation instructions before they are even able to build the docs!

As part of this change, I'm killing ChangeLog altogether and getting rid of
the roadmap from the top-level directory.

Monotone-Parent: f179dd0aec52a5faf255ec181b6a93dcefad2779
Monotone-Revision: f927003a9732d5c8b60af5f4fd9f9eb2976098ca

Monotone-Author: jmmv@NetBSD.org
Monotone-Date: 2009-10-24T18:29:30
Monotone-Branch: org.NetBSD.atf.src