Add for-each iteration to maps in atf-c
Modify the map module to support for-each iteration.  This requires providing
keys as part of the iterators, adding methods to get an iterator for the
beginning of the map, and adding the for-each macros.

This will be required to iterate over and dump the test case meta-data from
test programs so that atf-run can read it.

Monotone-Parent: 29743dbad6f59686374bb1a92b85897b8eb38839
Monotone-Revision: 36de2afdc6468225aa2a2f116f2265feb8dcbaef

Monotone-Author: jmmv@NetBSD.org
Monotone-Date: 2010-03-20T10:01:52