Change the behavior of developer mode
Developer mode used to be a collection of two related concepts: the
detection and enabling of multiple warning-detection flags, and the
enforcement of such flags plus the enforcement of assertions.

This change splits these two responsibilities: warning-detection flags
are now unconditionally used, and developer mode makes such warnings
be enforced.  The behavior of assertions is not changed.

Additionally, developer mode will only be enabled on repository
checkouts, not on formal releases.  The enforcement of warnings is a
constant source of pain for users, and the standard reply of "disable
developer mode" just does not help.

I am borrowing the code straight from the Kyua repository to implement
this.  The code used to be based on the one we had here, but it has
changed so much that it is just easier to copy it.

Monotone-Parent: e327a3367afd8c5dc604e18824d9846ff48e57b2
Monotone-Revision: 4b339688692a2e5ae8df382c60c57ce3f14f43f4

Monotone-Author: jmmv@NetBSD.org
Monotone-Date: 2011-08-14T19:55:26
Monotone-Branch: org.NetBSD.atf.src