Project hosting has moved to Google Code
Update links to the ATF web site as well as the documentation pointing
to the previous ad-hoc project hosting.  The project's web site,
downloads and source repository have moved to Google Code.  They are
not a project of its own though: the resources live within the Kyua
project.

Also change all the revision.h generation logic to account for Git
instead of Monotone so that the output of atf-version is meaningful
again.