Disable the build of the tools by default
This change adds a new --enable-tools flag to configure that allows
the user to request the build of the deprecated atf-run and atf-report
utilities.  Their build is disabled by default.

I am doing this right now because the new kyua-atf-compat package
provides compatibility scripts for these deprecated tools, but using
Kyua in the background.