Split contents of atf-c++'s utils
Move the noncopyable and auto_pointer classes defined in atf-c++'s utils
module to their own files (the former being public and the latter private).

This empties the utils module and leaves space to expose public utility
functions from this module, just like as we have done in the atf-c library.