Fix fork_test:stop in current versions of Linux
For some reason, the fork_test:stop test has stopped working in recent
versions of Linux, including Fedora 18 and Debian 7.0 testing.  The
spawned test program, upon reception of a CONT signal, does not really
continue.  Instead, it terminates with a strange error code (128 as
reported by bash), which does not appear to map to a specific signal.

Apply a super-ugly little hack to make the test pass and, while doing
so, remove some old baggage that dealt with test programs spawning test
cases.