Do not write unprintable characters in XML reports
Convert some unprintable characters to a *plain-text* representation in
XML reports.  This is a little hack but does the trick for now.

The NetBSD test suite has started spitting out a character with value 27,
and that confuses xsltproc when processing an XML-based report created by
atf-report.  Even converting the character to &#27; doesn't seem to work.