Flush atf-check messages
When a test case gets interrupted (due to a timeout), it can happen that
the stdout of the test does not include details about the last command
executed via atf_check because stdout was not flushed.  Flush stdout so
that we get a better idea of what was happening.

Requested in NetBSD PR bin/48284 by Andreas Gustafsson.