Fix various mandoc warnings
The Bd macro must specify one style of formatting for the block and
the blocks used to provide function definitions were not specifying
any.  Just added -ragged to them.

The Em macro behavior is cancelled if it is immediately followed by
another macro so just removed the Em macro in those cases.

Fixes issue 64