Allow overriding the path to the process_helpers
When locating the process_helpers program from the C++ tests, allow
overriding the path to the base directory where the atf-c tests live
via an ATF_C_TESTS_BASE define.

This is a workaround to allow the FreeBSD test suite to locate this
binary because the tests are installed (for good reason) using a
different layout than the one we provide and thus relative paths don't
work.