Stop using atf-version for the pkg-config tests
Add a couple of helper binaries whose purpose is to just print the
value of PACKAGE_VERSION.  Use these from the pkg-config tests to
query the built-in version value and compare it to the one provided
by the pkg-config files.

The reason I am adding two helper binaries is to avoid the nuisance
of having to cope with the FreeBSD test suite layout.  If we
attempted to use the version_helper binary of atf-c from the atf-c++
tests, we would have problems finding it.