Check if the compiler cannot detect unused tests
In the detect_unused_tests tests, check first to see if the compiler
will actually fail when a static global variable is declared from within
a macro.  If the compiler does not show a failure in that case (like it
happens with FreeBSD's clang 3.3), then set the test as an expected
failure and assume that the compiler is "broken" because our real tests
using the more complex atf headers will probably fail.

(Note that the compiler may not be strictly broken.  For example, it is
a design decision of clang to hide warnings risen by macros.  However,
it is indeed breakage from our point of view and I believe it's worth
signaling to the user.)