Move the parser module to a new utils library
Start decoupling code needed only by the tools from the libatf-c++
library by moving it to a new static-only utils library.

The goal is to make the tools not depend on libatf-c++ so that the
library can change at will without caring about the deprecated code.
The reason for this is that NetBSD will keep a code of the old
tools once we delete them and we don't want to break that copy by
mistake just by modifying the base libraries.

This change just adds the new utils library and moves the parser
module into it.  More will move later.