Move all tools-related code to a tools subdir
For simplicity reasons, contain all tools-related code into a tools
subdirectory.  This includes both helper code as well as the code of
the tools themselves.

This goes against the previous attempt of moving shared code into an
utils subdirectory and building a static library in it to be used by
the various tools.  Such approach was becoming too complex and hard
to handle.

The goal is to just dump any code in here, possibly duplicating what
still exists in atf-c/atf-c++, because we want this directory to not
depend on the shared atf libraries.  The tools directory will be
deleted on the next release anyway.