Decouple application module between tools and libs
Duplicate the application module from the atf-c++ library into the
tools directory so that we can simplify the version in atf-c++ by
removing everything not needed by the libraries.  In particular,
this really kills all useless screen wrapping for atf-sh and
atf-check which really is not a big deal.

As a result of this, we can also move the ui module altogether from
the libraries into the tools only.