Bump libatf-c++.so's major version
Theoretically, because only internal interfaces have been removed from
atf-c++ during the decoupling of the tools from the libraries, any test
programs should continue to work "just fine" with a newer version of
this library.  Only the tools would supposedly break, but they would be
rebuilt anyway as part of ATF's build.

That said, and because in practice things are not that simple, bump the
major version number of libatf-c++.so to be safe.