Use prebuilt Kyua toolchain if available
Attempt to download and unpack a prebuilt Kyua toolchain for Ubuntu
12.04 instead of rebuilding everything from scratch every time.  This
is to save time during build iterations and also to prevent flaky
results due to problems with the dependencies.

Not sure if this will work, but it seems to locally on a 12.04.4 VM.

Also install the built files under /usr/local instead of /usr and
make sure all build-time paths are correct.