Recommend the (expected, actual) idiom
This affects all the ATF macros and functions to compare values and is in
sync with what other testing libraries do.  Do this in the manual pages
and also rename the parameters in the header files (which has no effect
on the code but clarifies the syntax of the macros).

I'm not bothering to change any of the existing code to respect this style
for now given that the benefit is minimal.

Fixes #5.