Remove set -e from atf-sh
From the NEWS entry: This setting was initially added as a way to enable
a "strict" mode in the library and to make test cases fail fast when
they run unprotected commands.  However, doing so in the library is
surprising as the responsibility of enabling 'set -e' should be on the
user's code.  Also, 'set -e' introduces inconsistent behavior on
subshells and users do not expect that.

As part of this change, add a simple test to ensure that atf-sh continues
to work if the user decides to enable 'set -e'.

Fixes #14.