Release atf-0.21
Update release notes accordingly and bump the major version of the shared
libraries.

In theory, bumping the major version is not strictly necessary because we
only removed the internal config module... but there might be binaries out
there relying on it (see the old version of the ATF libraries in NetBSD).
Better be safe than sorry given that a symbol removal requires a major
version bump.