Prevent double-freeing a list object when managed
The 'managed' flag to a list indicates that its entries are dynamically
allocated and that the list is responsible for freeing them.  However,
a code path in the allocation of new nodes was not properly honoring
this.

atf_map uses a non-managed list and was triggering a double-free error
because it was explicitly deleting an entry on an error code path but
the list was mistakely also deleting it.

Fixes #23.