atf-c/atf-c++: use our new API to reset the resultsfile
All of the tests that fork_and_wait will fork(), fork() again, print some
stuff and exit, then the middle-fork parent will wait for the exit and do
some checks.  These checks are expected to fail in some cases, but what
we aren't expecting is for them to use the same resultfile (oops!).

Fix this by using our new API to reset the result file to /dev/null in the
middle-fork parent, so that our meta-tests are OK.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>