admin: fix CC/CXX issue in CI
The Cirrus environment may not have CC/CXX set, in which case we should
use some sensible default.  This causes problems as the compiled product
actually tries to use ${ATF_BUILD_CC} and ${ATF_BUILD_CXX}, and this fails
miserably as they're ultimately empty.

However, we also don't really need to do any of this if we don't have any
archflags to add, so just don't.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>