atf-check.1: stop with extraneous quoting
Quoted multi-word commands will be passed literally to exec rather than
split on arguments, which will simply not work.  Stop excessive quotes to
make sure we don't mislead folks into thinking they're needed, and slap -x
on the one example that needs it.

This is a follow up to PR #13.