Allow skipping a test that is expected to fail
I'm working on a test that will fail due to a kernel bug.  However, it
can't be executed unless certain hardware is present, and I can't check
if the hardware supports the test without running the entire test.  So
it makes sense to set atf_tc_expect_fail in the beginning and later to
atf_tc_skip if the hardware isn't suitable.  But ATF doesn't currently
allow that.