safe_main: refactor to avoid leaking tc memory on usage error
The function would previously leak memory on error in `safe_main(..)` as
the cleanup logic was previously only executed when an exception wasn't
thrown. This was because the testcase objects were being initialized,
but not cleaned up on exit.

Refactor the code so trivial usage errors will not leak memory on
error.

This helps eliminate some of the well-hit failure cases in the bootstrap
tests.

Signed-off-by: Enji Cooper <ngie@FreeBSD.org>