atf-check.cpp: remove unneeded copying into vector
atf::fs::path has a str() method, and it returns a std::string
constructor. I found this odd that a vector is used here to copy
std::string contents. As std::string is a typedef of
std::basic_string<char>, this additional copy is not needed.

Also, remove calling the str() method on error.

Signed-off-by: rilysh <nightquick@proton.me>