atf-check.cpp: remove unneeded copying into vector (#91)
atf::fs::path has a str() method, and it returns a std::string
constructor. I found this odd that a vector is used here to copy
std::string contents. As std::string is a typedef of
std::basic_string<char>, this additional copy is not needed.

Also, remove calling the str() method on error.

Notes by ngie@:

`buf.data()` returns `const char*` until C++17, so it can't be used in
combination with `mkstemp(..)` (it mutates the template argument).
Access the internal buffer with the `operator[]` instead.

Add the C++17+ form as well along with a TODO comment, so the code can
move to std::string::data eventually.

Co-authored by:	Enji Cooper <ngie@FreeBSD.org>

Signed-off-by: rilysh <nightquick@proton.me>
Co-authored-by: rilysh <nightquick@proton.me>