Build cleanup (#85)
* Garbage collect stray reference to auto_array

This entry should have been removed in the referenced commit.

Fixes:	6639d083e7d5b4b478397cc416cd42a756d17b0c

* Require C++-14 at bare minimum

This change imports ac_cxx_compile_stdcxx.m4 from gnu,org and makes use
of the `AX_CXX_COMPILE_STDCXX` macro in configure.ac to ensure that the
compiler specified supports C++-14 (at bare minimum). This is being done
to quell some issues reported by scan-build about the code using C++
range-based for-loops (a feature added in C++-11).

* Stop shipping atf-*-api(3) manpages

These manpages have been deprecated for at least a release. Remove them
and all of the logic associated with them.

Signed-off-by: Enji Cooper <ngie@FreeBSD.org>

* Expose WCOREDUMP(..) in a deterministic manner

WCOREDUMP is considered an extension to the POSIX spec on multiple
platforms, and thus is not automatically exposed on all platforms. Add the
relevant preprocessor defines to config.h via autoconf, then leverage them
in atf-c(3).

This helps ensure that the platforms which support WCOREDUMP properly
expose the macro.

Signed-off-by: Enji Cooper <ngie@FreeBSD.org>

* Add compiler feature detection for `__attribute__((nonnull))`

This feature is being detected so several functions can be appropriately
marked as taking non-NULL/-nullptr parameters, and the compiler and static
analyzers can (in turn) make intelligent decisions when optimizing and
analyzing code, respectively.

Signed-off-by: Enji Cooper <ngie@FreeBSD.org>

* Restore AC_PROG_CPP

This was accidentally deleted post-0.21 release. It's still needed by
tests and build infrastructure.

Signed-off-by: Enji Cooper <ngie@FreeBSD.org>

* atf-c/detail/env.c: remove unnecessary complexity

The libcalls used have been in the POSIX standard since 2008.1. Require
them to be present instead of conditionally hoping they're present.

This fixes an issue where the autoconf code was messing up with a
combination of clang tools, which resulted in the autoconf code failing
to properly determine whether or not the functions were available.

Signed-off-by: Enji Cooper <ngie@FreeBSD.org>

* Remove ATF_BUILD_CXX require.progs check

This particular check unfortunately doesn't work when ATF_BUILD_CXX
contains multiple CXXFLAGS, or contains a path with spaces in it. Remove
this check to unbreak the dependent tests
post-793d4640031dc06ce8a239ffa9ab61322104c4ca.

Signed-off-by: Enji Cooper <ngie@FreeBSD.org>

* Remove redundant C++ toolchain check

The C++ toolchain is sanity checked when C++14 conformance is checked;
there's no reason why we need to check whether or not the C++ toolchain
works again.

Signed-off-by: Enji Cooper <ngie@FreeBSD.org>

---------

Signed-off-by: Enji Cooper <ngie@FreeBSD.org>