dynstr_test: fix build with system clang on macOS
The system version of clang on macOS rightfully flags defining a
`const size_t` and using it when defining an array as a GNU extension
for the standard and the default warning set picked by the Apple in
the compiler. Replace the `const size_t` with a more standard C constant
via `#define`.

Similar things can be achieved using a newer C standard (C11 seems to
require this support as a non-compiler extension), but it's simpler
solving the issue this way instead of forcing all of the consumers to use
a newer C standard. Requiring a newer standard would require changes in
the FreeBSD build process as it doesn't use autotools when building the
project, and potentially other distros, and it would increase the amount
of required testing.

Signed-off-by: Enji Cooper <ngie@FreeBSD.org>